fix ofHashMap
