--- java/fury-core/src/test/java/io/fury/collection/MapsTest.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2023 The Fury Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.fury.collection;
-
-import static org.testng.Assert.*;
-
-import com.google.common.collect.ImmutableMap;
-import java.util.HashMap;
-import org.testng.annotations.Test;
-
-public class MapsTest {
-
-  @Test
-  public void testOfHashMap() {
-    assertEquals(Maps.ofHashMap(1, 2), new HashMap<>(ImmutableMap.of(1, 2)));
-  }
-}
